Initial version of a37x0 utilities
This patch added Armada37x0's utility tools which used by
the bootloader, e.g., u-boot or ATF.

The TBB tool is used to generate the boot image for both
trusted and non-trusted boot which compiles with Marvell
Wireless Trusted Platform (WTP). The source code of this
tool is located in the WTP Toolkit package (wtptp_tool)
direcotry - 'src/TBB_Linux'. For more information about
the required library and compiler, please refer to the
'readme.txt'. The executable tool file is also provided
in 'linux/TBB_Linux'.

The TBB tool requires the specific profile to describe
the structure of the boot image components as its input
in order to generate the boot image header (TIM Header).

The utility tools includes the TIM build scripts and the
templates for TIM Header, which is located in 'trusted'
and 'untrusted'. The boot image header also contains a
set of reserved area packages:
- Clock package: clocks-CPU-XXX-DDR-YYY.txt
- DDR package: DDR<type>-<speed_bin>-<cs_num>.txt
- GPP package: gpp1.txt and gpp2.txt

DDR porting is based on the predefined topology files -
DDR_TOPOLOGY_<N>.

For trusted boot, OEM Keys (CSK, KAK, etc) are defined
in 'keys' directory.

In addition, the utility tool includes the firmware to
burn the efuse registers for the trusted boot, which
is running on CM3 processor. The CM3 firmware is part
of the boot image as WTMI component. The source code is
located in 'wtmi' directory.

Change-Id: I6e335445db131c69e0b120c2033589c769985643