tool: add WTP download tool
The WTP downloader tool is a Linux version tool for
downloading boot image for UART booting in order to
recover the system with a mal-functional bootloader
image in the on-board flash device.

The tool can be also used to trap BootROM into UART
R/W mode or WTP download mode with the escape sequence.

This patch submits the source code of this tool along
with the pre-built binary (initial verison 3.3.11.15).

    - wtptp_tool/linux/WtpDownload_linux

A README file is provided to explain the typical
usage for the tool and a simple how-to about building
the tool by yourself.

Change-Id: Ic355bc1c2de6e7924407338a969136f50f9bef16
Signed-off-by: Wilson Ding <dingwei@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/41724
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Neta Zur Hershkovits <neta@marvell.com>