ddr: add wtmi_ddr module
The WTMI DDR module is supposed to include the entire DDR PHY
training flow. As the first stage, only DLL tuning is supported.
Prior to DDR PHY training, WMTI sets DDR to self-refresh mode
and then raises the DDR clock to the target frequecy. So that
the bootloader image can be retained in DRAM. The system
clock initialization and AVS setting are completed along with
this step. Once DDR PHY training is completed successfully,
WTMI will kick off the application processor to proceed the boot.
Thus, WTMI DDR module introduces the following components to
achieve the above boot flow:
   ddr_main:   main entry to call ddr functions
   clock:      APIs to do clock initialization
   avs:        APIs for AVS configuration
   apctl:      kick the AP start via mailbox
   ddrcore:    APIs for DDR init and DLL training.

Change-Id: Ib096e965f67f105d6f6864e4a303a0acc09e1368
Signed-off-by: Terry Zhou <bjzhou@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/42573
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Wilson Ding <dingwei@marvell.com>
Reviewed-by: Hua Jing <jinghua@marvell.com>