ddr: move ddr phy training to wtmi
The reason for this transition is to "enhance DDR PHY parameters
tuning, which TIM can only do very limited training and consumes
a lot of time", we will move the DDR training, system clock
initialization and AVS setting to wtmi, in TIM, we will only
initialize the DDR related clock and set it to a very LOW
frequency, thus, the DDR chip can be set into self-refresh mode
in wtmi and change to a higher destination frequency. Dues to
this update (change clock settings in wtmi), BootROM is not'
supposed to kickoff the application processor as usual. Instead,
it will be done in WTMI after DDR finished PHY training at the
target frequency.

Change the Makefile to support this feature.
After this patch, we will got the update below:
   Set DDR to LOW frequency in TIM (DDR3-300M, DDR4-400M)
   Move DLL tuning from TIM to WTMI
   Kickoff AP and update AVS settings in wtmi
   Build the wtmi_ddr as a static lib and call from wtmi entry

Change-Id: Ie0b42a206fa0cecbcc7a3d269ea9db19e485d0c0
Signed-off-by: Terry Zhou <bjzhou@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/42575
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Wilson Ding <dingwei@marvell.com>
Reviewed-by: Hua Jing <jinghua@marvell.com>