wtmi: reorganize source tree to support cm3 freeRTOS image
Now, CM3's typical application is fuse programming or FreeRTOS. In
addition to the runtime applications, CM3 is responsible for system
early initialization (before kicking off theAP core). It includes AVS
setting, clock-tree setup and DDR PHY dynamic training. The sys-init
function should be applied to all kinds of CM3 applications.

The sys-init fucntion used to be built as part of fuse application.
This patch disconnected it from the specific application. The sys-init
function becomes an independent binary extended to CM3 applications.
Thus, WTMI image combines multiple binaries - runtime application
image (a.k.a., fuse application or FreeRTOS) followed by sys-init
image.

The new WTMI image structure also change the boot sequence of CM3 and
image exec_addr. The load_addr and exec_addr of WTMI image used to be
the same - 0x1FFF_FFFF. The multi-image is supposed to be loaded to
the same load_addr as is. However, CM3 starts from the sys-init image.
Once sys-init image is executed, it will jump to the entry point of the
application image, i.e., 0x1FFF_FFFF. From then, sys-init image is not
used anymore and will be wiped out from the memory.

The following is the major changes to the source tree structure.
- Move the folders in 'ddr' directory to the correlated directories of
  the boot image components:
      ddr/wtmi-ddr => wtmi/sys-init
      ddr/tim-ddr  => tim/ddr
- Move the fuse application related code to wtmi/fuse directory
- Exact the common files of fuse applicaton and sys-init function into
  a new folder - 'wtmi/common'

Change-Id: I9982760f0d0e1395d4a7e0d14a5c9ce009f58926
Signed-off-by: Ken Ma <make@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/45445
Reviewed-by: Wilson Ding <dingwei@marvell.com>
Tested-by: Wilson Ding <dingwei@marvell.com>
Reviewed-by: Hua Jing <jinghua@marvell.com>