wtmi: Calculate DRAM win settings upon memory size
CM3 has two address decoding windows, which can
be linear mapped to the contineous memory base.
It could translate up to 1.5GB memory address.

The patch removed the DRAM window bonding to per
each DDR topology. The window is supposed to be
configured according to the chip size.
- With the DDR chip of small capacity (< 1GB),
the first window can translate the address for
both chips. Considering DDR training requires
only a small piece of memory, the access to the
the second chip is limited to 512MiB at maximum.
- With the DDR chip of large capacity (> 1GB),
the second window must be used and remapped to
the base address of the second chip. It allows
to access the first chip for the first 1GB,
while access the second chip for the first 512MB.

This patch also added the dumps for the window
settings and available memory region for memory
test.

Change-Id: I140423088f078cab6dc59e39e8a0ce5e0265d551
Signed-off-by: Wilson Ding <dingwei@marvell.com>
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hua Jing <jinghua@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/49765
Tested-by: Hua Jing <jinghua@marvell.com>