wtmi: Add autoconf.h and filechk rule in Makefile
This patch set the dependencies of wtmi image to
a few build options, which WTMI code is not aware
of the changes with incremental build.

- General build flags, such as clock preset.
- DDR topology file: The DDR topology parameters
is described in a plain text file, which is also
pass to WTMI code as a set of build flags.

All these build options will be dumped to a header
file (autocfg.h), which is generated and tracked
by Makefile. With the new header file, WTMI DDR
code removes the bonding to the DDR topology ID.

The DDR log was slightly improved in this patch
accordingly.

Change-Id: Ib2150b27c272c7ec1645f78a0de0e0972f8105da
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Hua Jing <jinghua@marvell.com>
Signed-off-by: Allen Yan <yanwei@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/49766
Tested-by: Hua Jing <jinghua@marvell.com>