makefile: Add inline build for ddr tool
This patch removed the pre-built binary of ddr tool in
A3700_utils source tree. Instead, the tool is supposed
to be built with the latest mv_ddr code. So A3700_utils
can use the most updated tool to generate the TIM DDR
code.

In order to build the tool, mv_ddr path must be either
exported as a build environment or specified in 'make'
command. As a sub-directory to ATF makefile, the path
is supposed to be passed along with ATF build.

Since ddr tool was upgraded to the latest version, the
porting guide and DDR topology files have been updated
accordingly. The DDR TIM is now generated as one of the
step in TIM creation procedure by buildtim.sh script.

Change-Id: I0bd9711a8d5bc08ce7b4532fa6f592c7adc1f624
Signed-off-by: zachary zhang <zhangzg@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/47401
Reviewed-by: Wilson Ding <dingwei@marvell.com>
Tested-by: Wilson Ding <dingwei@marvell.com>