tim: Print mv_ddr version and configuration on UART
mv_ddr code is included from external repository. This change adds into
GPP1 section simple code which prints mv_ddr version and configuration over
UART, like it done for printing TIM version.

This would allow to identify which mv_ddr release or git commit was used
for building wtmi/tim binary. Also it would allow to identify which ddr
configuration (DDR_TOPOLOGY) was used for building of ddr part of wtmi.

Version and configuration string is printed only in the case when boot
image is not for UART as it would break UART downloading.

Signed-off-by: Pali Rohár <pali@kernel.org>